requirements.txt
openai>=1.0.0
playwright>=1.30.0
python-dotenv>=1.0.0
gradio>=3.0.0
Pillow>=9.0.0